Improved navigation bar
Added a bar with the width of 0 underneath each link. If the link is hovered, the bar's width will be 100%. This is a nice indication for the user that they are hovering over the link. I also made some adjustments to the text, button and bar sizes of the header.